requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.1
requests==2.32.3
Flask==3.0.3
